서버, 클라이언트 통신 구현 - 코딩만 하고 아직 테스트는 하지 않음

+++ taxi-game-3d-server/TaxiGame3D.Server/Templates/StageTemplate.cs
@@ -9,5 +9,5 @@ public class StageTemplate
     public double FareRate { get; set; }
 
     [JsonIgnore]
-    public double MaxCoin => Distance * FareRate;
+    public long MaxCoin => (long)(Distance * FareRate);
 }
